[ refactor ] 컨밴션에 따른 UpdateReservationMessageDTOV1 구조 수정 ( #17 )
